\name{data_HS_1939}
\alias{data_HS_1939}
\docType{data}
\title{Holzinger and Swineford (1939) data for 9 mental ability test scores}
\description{Raw data for 9 mental ability test scores in 301 seventh- and 
eighth-grade children, from two schools. The data were provided by Holzinger 
and Swineford (1939), adapted from the MBESS package.

test1 = Visual perception

test2 = Cubes

test3 = Lozenges

test4 = Paragraph comprehension

test5 = Sentence completion

test6 = Word meaning

test7 = Speeded addition

test8 = Speeded counting of dots

test9 = Speeded discrimination straight and curved capitals

school = 'Grant-White' or 'Pasteur'
}
\usage{data(data_HS_1939)}
\references{Holzinger, K., and Swineford, F. (1939). A study in 
factor analysis: The stability of a bifactor solution. 
\emph{Supplementary Educational Monograph, no. 48.} 
Chicago: University of Chicago Press.}
\examples{
head(data_HS_1939)
}




