\name{data_RSE}
\alias{data_RSE}
\docType{data}
\title{Item-level dataset for the Rosenberg Self-Esteem scale}
\description{A data frame with 800 observations on the 10 items from the 
Rosenberg Self-Esteem scale. Negatively-worded items were reverse-coded.}
\usage{data(data_RSE)}
\references{Rosenberg, M. (1965). \emph{Society and the adolescent self-image.} 
Princeton University Press.}
\examples{
head(data_RSE)
\donttest{
}
}
