\name{data_RSE_not_recoded}
\alias{data_RSE_not_recoded}
\docType{data}
\title{Item-level dataset for the Rosenberg Self-Esteem scale}
\description{A data frame with 300 observations on the 10 items from the 
Rosenberg Self-Esteem scale. Negatively-worded items are not reverse-coded.}
\usage{data(data_RSE_not_recoded)}
\references{Rosenberg, M. (1965). \emph{Society and the adolescent self-image.} 
Princeton University Press.}
\examples{
head(data_RSE_not_recoded)
\donttest{
}
}
