\name{data_RSE_sex}
\alias{data_RSE_sex}
\docType{data}
\title{Item-level dataset for the Rosenberg Self-Esteem scale plus gender}
\description{A data frame with 400 observations for males and 400 observations
for females on the 10 items from the Rosenberg Self-Esteem scale.
Negatively-worded items were reverse-coded.}
\usage{data(data_RSE_sex)}
\references{Rosenberg, M. (1965). \emph{Society and the adolescent self-image.} 
Princeton University Press.}
\examples{
head(data_RSE_sex)
\donttest{
}
}
