\name{data_SDT}
\alias{data_SDT}
\docType{data}
\title{data_SDT}
\description{A data frame with scores for 366 cases on 12 variables that have  
	the same correlational structure as those reported by Snchez-Oliva et al. (2017).
	The data were generated using the values in their Table 1, p. 179.
	There are four scores for each of three needs, pertaining to autonomy, 
	competence, and relatedness at work.
}
\usage{data(data_SDT)}
\references{Snchez-Oliva, D., Morin, A. J. S., Teixeira, P. J., Carraa, E. V., 
Palmeira, A. L., & Silva, M. N. (2017). A bifactor exploratory structural 
equation modeling representation of the structure of the 
Basic Psychological Needs at Work Scale. 
\emph{Journal of Vocational Behavior, 98,} 173187.}
\examples{
head(data_SDT)
}





