\name{EFA.dimensions-defunct}
\alias{PA_FA}
\alias{MAXLIKE_FA}
\alias{IMAGE_FA}
\title{Defunct functions}
\description{
  The functions below are defunct and have been removed and replaced.
}
\details{
\itemize{
    \item \code{PA_FA} has been replaced by \code{\link{EFA}}
    \item \code{MAXLIKE_FA} has been replaced by \code{\link{EFA}}
    \item \code{IMAGE_FA} has been replaced by \code{\link{EFA}}
}
}
\author{Brian P. O'Connor \email{brian.oconnor@ubc.ca}}
