\name{EFA.dimensions-package}

\alias{EFA.dimensions-package}

\docType{package}

\title{EFA.dimensions}

\description{This package provides exploratory factor analysis-related functions for 
	assessing dimensionality. There are functions for seven different procedures 
	for determining the number of factors, including functions for parallel analysis 
	and the minimum average partial test. There are functions for conducting principal 
	components analysis, principal axis factor analysis, maximum likelihood factor 
	analysis, image factor analysis, and extension factor analysis, all of which can 
	take raw data or correlation matrices as input and with options for conducting 
	the analyses using Pearson correlations, Kendall correlations, Spearman correlations, 
	or polychoric correlations. Varimax rotation, promax rotation, and Procrustes 
	rotations can be performed. Additional functions focus on the factorability of 
	a correlation matrix, the congruences between factors from different datasets, and 
	for assessing local independence. 
}

\references{
{O'Connor, B. P. (2000). SPSS and SAS programs for determining the number of components 
 using parallel analysis and Velicer's MAP test. \emph{Behavior Research Methods, Instrumentation, 
 and Computers, 32,} 396-402. doi:10.3758/bf03200807}
\cr\cr {O'Connor, B. P. (2001). EXTENSION: SAS, SPSS, and MATLAB programs for extension analysis. 
 \emph{Applied Psychological Measurement, 25,} p. 88. doi:10.1177/01466216010251011.}
\cr\cr {Fabrigar, L. R., & Wegener, D. T. (2012). \emph{Exploratory factor analysis.} 
  New York, NY: Oxford UNiversity Press. ISBN:978-0-19-973417-7}
\cr\cr {Field, A., Miles, J., & Field, Z. (2012). \emph{Discovering statistics using R.}
  Los Angeles, CA: Sage. ISBN:978-1-4462-0045-2}
}
